requirements.txt
black==22.3.0
blinker==1.4
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Mail==0.9.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
pathspec==0.9.0
platformdirs==2.5.1
protobuf==3.19.4
tomli==2.0.1
Werkzeug==2.0.3
python-dotenv==0.20.0

